Added a rule to exclude to-test.md from deletion.

+++ tools/build-jetpack.sh
@@ -68,6 +68,9 @@ NODE_ENV=production BABEL_ENV=production $TARGET_DIR/node_modules/.bin/gulp
 echo "Purging paths included in .svnignore, .gitignore and .git itself"
 # check .svnignore
 for file in $( cat "$TARGET_DIR/.svnignore" 2>/dev/null ); do
+    if [[ $file == "to-test.md" ]]; then
+        continue
+    fi
     rm -rf $TARGET_DIR/$file
 done
 echo "Done!"
